requirements.txt
# Install PyTorch with CUDA 12.6
--extra-index-url https://download.pytorch.org/whl/cu126
numpy==1.25.0
torch
torchvision
torchaudio
# Install NuScenes devkit
nuscenes-devkit
diffusers==0.26.3
safetensors==0.5.3
transformers==4.31.0
plotly==6.0.1
einops
huggingface-hub==0.21.3